requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
uvloop==0.20.0
qrcode[pil]==7.4.2
Pillow==10.4.0
requests==2.32.3
openai>=1.0.0
# тесты (опционально: pytest tests/)
pytest==8.3.3
pytest-asyncio==0.24.0
